Improve timer start time handling
The time at which a timer was started had been hard coded to 'now' meaning the timer would fire immediately for the first time. This can now be set individually when calling start() on the timer and is also maintained more consistently when changing the interval with the timer still running.